creating test namespace
